scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=9.0.0
numpy>=1.24.0
